pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "numerology-map"
version = "0.1.0"
description = "Generates numerology map."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "Authlib==1.6.9",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.5",
    "click==8.3.2",
    "cryptography==46.0.5",
    "fastapi==0.135.1",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "idna==3.11",
    "Mako==1.3.10",
    "MarkupSafe==3.0.3",
    "psycopg2==2.9.11",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "python-decouple==3.8",
    "python-decouple-stubs==3.8.0.4",
    "python-dotenv==1.2.2",
    "requests==2.32.5",
    "SQLAlchemy==2.0.48",
    "starlette==0.52.1",
    "stripe==14.4.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "urllib3==2.6.3",
    "uvicorn==0.44.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]